finish refactor for row and column blocks. relates to #504.
