+++ src/blocks/row/column/inspector.js
@@ -3,8 +3,8 @@
  * Internal dependencies
  */
 import applyWithColors from './colors';
-import { BackgroundPanel } from '../../../../components/background';
-import DimensionsControl from '../../../../components/dimensions-control/';
+import { BackgroundPanel } from '../../../components/background';
+import DimensionsControl from '../../../components/dimensions-control';
 
 /**
  * WordPress dependencies
