+++ src/blocks/row/index.js
@@ -3,46 +3,20 @@
  */
 import './styles/editor.scss';
 import './styles/style.scss';
-import { Deprecation, DeprecationVideo } from './components/deprecations';
-import Edit from './components/edit';
+import edit from './edit';
+import transforms from './transforms';
+import deprecated from './deprecated';
+import save from './save';
+import metadata from './block.json';
+import getEditWrapperProps from './getEditWrapperProps';
 import icons from './../../utils/icons';
 import { BackgroundAttributes } from '../../components/background';
 import DimensionsAttributes from '../../components/dimensions-control/attributes';
-import Save from './components/save';
-import metadata from './block.json';
 
 /**
  * WordPress dependencies
  */
 const { __ } = wp.i18n;
-const { createBlock } = wp.blocks;
-
-/**
- * Generates a layout based on the :row prefix.
- * The number of :: represents the number of columns to input.
- * We fallback to the standard divided column layouts.
- *
- * @param {int} columns The number of columns.
- * @return {String} defaultLayout
- */
-function generateLayout( columns ) {
-	let defaultLayout;
-
-	switch ( columns ) {
-		default:
-		case 2:
-			defaultLayout = '50-50';
-			break;
-		case 3:
-			defaultLayout = '33-33-33';
-			break;
-		case 4:
-			defaultLayout = '25-25-25-25';
-			break;
-	}
-
-	return defaultLayout;
-}
 
 /**
  * Block constants
@@ -76,66 +50,15 @@ const settings = {
 		coBlocksSpacing: true,
 	},
 
-	transforms: {
-		from: [
-			{
-				type: 'prefix',
-				prefix: ':row',
-				transform: function( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-					} );
-				},
-			},
-			...[ 2, 3, 4 ].map( ( columns ) => ( {
-				type: 'prefix',
-				prefix: Array( columns + 1 ).join( ':' ) + 'row',
-				transform( content ) {
-					return createBlock( `coblocks/${ name }`, {
-						content,
-						columns,
-						layout: generateLayout( columns ),
-					} );
-				},
-			} ) ),
-		],
-	},
+	transforms,
 
-	edit: Edit,
+	edit,
 
-	getEditWrapperProps( attributes ) {
-		const { id, layout, columns, hasAlignmentControls } = attributes;
+	getEditWrapperProps,
 
-		// If no layout is seleted, return the following.
-		if ( ! layout ) {
-			return { 'data-id': id, 'data-columns': columns, 'data-layout': 'none' };
-		}
+	save,
 
-		if ( hasAlignmentControls === false ) {
-			return { 'data-align': '' };
-		}
-
-		return { 'data-id': id, 'data-columns': columns, 'data-layout': layout };
-	},
-
-	save: Save,
-
-	deprecated: [
-		{
-			attributes,
-			save: DeprecationVideo,
-		},
-		{
-			attributes: {
-				isStackedOnMobile: {
-					type: 'boolean',
-					default: true,
-				},
-				...attributes,
-			},
-			save: Deprecation,
-		},
-	],
+	deprecated,
 };
 
 export { name, title, icon, settings };
